+++ src/main/java/com/team/cafe/keyword/KeywordType.java
@@ -8,13 +8,13 @@ import java.util.ArrayList;
 import java.util.List;
 
 @Entity
-@Table(name = "keyword_type",
+@Table(name = "keyword_type", // 인덱스용 !
         uniqueConstraints = @UniqueConstraint(name="uk_keyword_type_name", columnNames = "type_name"))
 @Getter
 @Setter
 public class KeywordType {
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @Column(name="type_name", length=30, nullable=false)
     private String typeName; // persona / space / menu 등
